RFC: Add a mechanism to avoid including unnecessary files in the trace
This is more a kick off point for discussion than a serious PR at this point,
but I figure it's always better to discuss with working code than in the abstract.

We're planning to start collecting large numbers of rr traces from failing CI jobs.
However, at the moment, these traces are a bit large for comfort. However, they
don't need to be that large, because a good fraction of their size comes from
binaries that we're already storing on our artifact server, and thus don't need
to store separately in the trace. This PR adds an `--index-dir` option to `rr pack`.
rr will index each such directory, cataloguing all contained files by content hash.
It will then symlink the index directory into the trace, rewrite the mmap entries
to point into this symlink and as usual delete any unreferenced files. The idea is
to send this directory over to the replay machine, rewrite the symlink to a directory
containing the same immutable binaries that were in the index directory and replay
the trace that way.

This works ok, and is probably what we'll do for the short term, but it's not a
satisfying long-term solution, since we don't record any information whatsoever
about what was in the index directory. I'm thinking a better solution would be
to have a custom `external` record type that we could use with a custom packer
to record exactly where on our artifact server the relevant binary can be found
(which is easy, since those binaries are already content-addressed). That said,
the effort to implement that seemed not-insubstantial, so this is the poor man's
version of that.

If the version in this PR is intersting, I'm happy to clean it up and get it in
properly, but as I said, I'd like to personally move away from it in the medium
term, so I don't need it merged.