[GdbServer] Look for missing file in the ld rendezvous
The --serve-files option I recently added is very useful for
debugging traces from a different machine, where the original
file system may not be locally available for rr to read symbols
from (similarly if the local file system has been deleted or
the recording is from the inside of a container that the replayer
may not be able to see). However, for dynamic libraries opened
via symlink, we currently fail to serve them using this mechanism,
since we nowhere record the correspondence between the name of the
symlink and the name of the underlying file on the record side.
GDB however, is getting the original symlink paths
from the ld rendezvous struct.
This patch is a slightly hacky way to still serve
those files: If we didn't find the correct path in our memory
image, we too look for the ld rendezvous struct and see if the
filename came from any of those entries. If so, it's easy to
look at the memory map corresponding to this library to serve
the corresponding file. This does go a little further into
debugger frontend territory than I'd otherwise like to see rr
go and of course there's various corner cases that this doesn't
handle right, but the GDB frontend is still most people's first
exposure to rr, so I'd like to make it as functional as possible
and in practice this seems to work nicely for all the traces I've
tried.