Try to fix up x7 before processing a syscall exit
During a syscall stop, the kernel lies about x7 and ignores writes to
it. That's usually not a problem as long as we don't look at it. However,
it becomes a problem if we do any AutoRemoteSyscall'ing during the exit
processing. AutoRemoteSyscalls might leave us in either a singlestep
or a syscall exit trap. If the former, then trying to re-apply the
registers that it thought the task had before the AutoRemoteSyscall
will actually end up modifying the real x7 register (because we're
no longer in a syscall stop where the kernel would ignore such a
write). This is all a huge mess and I have an ongoing thread with
the kernel maintainers pleading for an option to give some sanity
here. That said, this seems to work in the meantime.