More robust exits from coredumping signals
Coredumping signals are extremely annoying. Every task in the
same address space, save for the one doing the coredumping
needs to reach its exit event before that task can proceed.
However, we must make sure not to accidentally continue a
task past its exit state, because otherwise it too will hang
until all other tasks have continued. Unfortunately, we don't
have a "ptrace-continue-unless-in-exit-state" command, so we
try to waitpid beforehand to make sure we're not in this
situation. That seems to work reasonably well on x86, but
for some reason not on aarch64, where I'm routinely observing
the waitpid not reportingt hat the task has stopped, but
the subsequent ptrace continue, moving it right past the stop.
This change tries to make that race detection slightly less
necessary by not provoking it during coredumping signals
that we inject. Instead, we mark that we're expecting a
ptrace exit event for every task that's about to die and
then refuse to schedule these tasks even if they otherwise
look runnable. This may feel like the return of unstable_exit,
but note that once the exit event arrives, we still process
the exits in an orderly manner, recording each tasks robust
futex changes etc. in order, so we don't lose any of the
advantages that removing unstable exit gave us. Instead,
we're simply avoiding the above described race, and relying
on the existing logic to handle everything properly.