Clear a pending group stop upon receiving SIGCONT
We may not have applied the group stop yet, but we definitely
need to clear it the task sending it to us doesn't know we
delayed it. This happens consistently during the exec_stopsig
test on aarch64, but I suspect it could easily happen elsewhere
and just happened to manifest here.