RFC: Pretend to the tracee that librrpage is the vdso
We were runnign into another #2721 like situation, except this time
due the vdso patch (for which we deliberately destroy the unwind info
to avoid confusing the unwinder - unfortunately, we'd not only like it
to not be confused, we'd also like it to work). To solve this, I
propose to stop patching the vdso alltogether and instead implement
a VDSO-compatible ABI in librrpage that can then a-priori do
whatever we need it to while simultaneously having correct unwind
information. As a bonus, this also lets us rip out the heuristic for
mapping librrpage, since the dynamic linker is happy to take what
it thinks of as the VDSO as a pre-relocated object in memory.

This seems to work fairly well and results in a bunch of nice simplification
of the Monkeypatcher code. There are two test failures because GDB
seems to have some special ideas about how to unwind the vdso (because
of course it does), so I need to look into how to convince GDB to
do the right thing, but before I do that, I wanted to check on
the direction here. I quite like it, but I'm wondering if it's a bit
too cute. Any concerns?