[rrpage]: Set `max-page-size` to `0x1000`
This prevents systems with huge pages by default causing the alignment
of program segments to become so large that the special placement of
`rrpage` segments gets borked.  We set a default of 4KB page size as a
compatible default for all `x86_64` arches, once other architectures are
supported this may need to become more flexible.