Detach: Unpatch syscalls before detaching
When detaching we leave the preload library in the former tracee.
We do lock the syscallbuf, so the tracee shouldn't really make use
of it, but I neglected to consider that the tracee might launch
additional threads, which, if they used a previously patched syscall,
would then share the syscallbuf stack. I considered simply setting
the stack nesting level to 1, which would force the syscallbuf
back onto the original stack, but this doesn't work because we
still use the preload thread locals for thread space and only
one thread may use these at a time. Additionally, it the tracee
may not expect to have its stack used in this way, which is why
we're switching stacks in the first place. This instead tries to
prevent the tracee from using the preload library entirely by
unpatching the syscalls. I considered trying to unpatch the
syscalls themselves, but I was worried that the tracee may
have modified said memory (e.g. because it was JITTed, or
because the tracee wants to set its own breakpoints, etc),
and detecting such conditions seemed very complex. Instead,
this unpatches the extended jump pages, which the tracee
should never really touch. Additionally, on x86, we need
to unpatch the vsyscall. The same JIT concerns don't really
apply here, so I'm just unpatching it directly. However,
I'm also expecting to just get rid of this patching in #2772,
so this code path is probably not particularly long-lived.

Ref https://github.com/JuliaLang/julia/issues/39068